functions/gmail_listener/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-secret-manager
google-cloud-functions
google-cloud-storage
google-cloud-core
google-cloud-pubsub
pydantic
google-genai